// src/app/models/restaurant.ts
export interface Restaurant {
    id: number;
    note: number;
    appreciation: number;
    prixmoyen: number;
    adresse: string;
    ville: string;
    datecreation: Date;
    nom: string;
    quartierid: number;
    categorieid: number;
    notecopiosite: number;
    notedeliciosite: number;
    notecadre: number;
    noteaccueil: number;
}